refactor: use stream reader to re-use stream code
