+++ chat_stream.go
@@ -2,12 +2,8 @@ package openai
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
-	"io"
-	"net/http"
 )
 
 type ChatCompletionStreamChoiceDelta struct {
@@ -31,52 +27,12 @@ type ChatCompletionStreamResponse struct {
 // ChatCompletionStream
 // Note: Perhaps it is more elegant to abstract Stream using generics.
 type ChatCompletionStream struct {
-	emptyMessagesLimit uint
-	isFinished         bool
-
-	reader         *bufio.Reader
-	response       *http.Response
-	errAccumulator errorAccumulator
+	*streamReader
 }
 
-//nolint:dupl // stream repetition code needs to be optimized
 func (stream *ChatCompletionStream) Recv() (response ChatCompletionStreamResponse, err error) {
-	if stream.isFinished {
-		err = io.EOF
-		return
-	}
-
-	var emptyMessagesCount uint
-
-waitForData:
-	line, err := stream.reader.ReadBytes('\n')
+	line, err := stream.streamReader.Recv()
 	if err != nil {
-		if errRes, _ := stream.errAccumulator.unmarshalError(); errRes != nil {
-			err = fmt.Errorf("error, %w", errRes.Error)
-		}
-		return
-	}
-
-	var headerData = []byte("data: ")
-	line = bytes.TrimSpace(line)
-	if !bytes.HasPrefix(line, headerData) {
-		if writeErr := stream.errAccumulator.write(line); writeErr != nil {
-			err = writeErr
-			return
-		}
-		emptyMessagesCount++
-		if emptyMessagesCount > stream.emptyMessagesLimit {
-			err = ErrTooManyEmptyStreamMessages
-			return
-		}
-
-		goto waitForData
-	}
-
-	line = bytes.TrimPrefix(line, headerData)
-	if string(line) == "[DONE]" {
-		stream.isFinished = true
-		err = io.EOF
 		return
 	}
 
@@ -85,7 +41,7 @@ waitForData:
 }
 
 func (stream *ChatCompletionStream) Close() {
-	stream.response.Body.Close()
+	stream.streamReader.Close()
 }
 
 // CreateChatCompletionStream — API call to create a chat completion w/ streaming
@@ -108,10 +64,12 @@ func (c *Client) CreateChatCompletionStream(
 	}
 
 	stream = &ChatCompletionStream{
-		emptyMessagesLimit: c.config.EmptyMessagesLimit,
-		reader:             bufio.NewReader(resp.Body),
-		response:           resp,
-		errAccumulator:     newErrorAccumulator(),
+		streamReader: &streamReader{
+			emptyMessagesLimit: c.config.EmptyMessagesLimit,
+			reader:             bufio.NewReader(resp.Body),
+			response:           resp,
+			errAccumulator:     newErrorAccumulator(),
+		},
 	}
 	return
 }
