+++ stream.go
@@ -2,13 +2,9 @@ package openai
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
-	"net/http"
 )
 
 var (
@@ -16,52 +12,12 @@ var (
 )
 
 type CompletionStream struct {
-	emptyMessagesLimit uint
-	isFinished         bool
-
-	reader         *bufio.Reader
-	response       *http.Response
-	errAccumulator errorAccumulator
+	*streamReader
 }
 
-//nolint:dupl // stream repetition code needs to be optimized
 func (stream *CompletionStream) Recv() (response CompletionResponse, err error) {
-	if stream.isFinished {
-		err = io.EOF
-		return
-	}
-
-	var emptyMessagesCount uint
-
-waitForData:
-	line, err := stream.reader.ReadBytes('\n')
+	line, err := stream.streamReader.Recv()
 	if err != nil {
-		if errRes, _ := stream.errAccumulator.unmarshalError(); errRes != nil {
-			err = fmt.Errorf("error, %w", errRes.Error)
-		}
-		return
-	}
-
-	var headerData = []byte("data: ")
-	line = bytes.TrimSpace(line)
-	if !bytes.HasPrefix(line, headerData) {
-		if writeErr := stream.errAccumulator.write(line); writeErr != nil {
-			err = writeErr
-			return
-		}
-		emptyMessagesCount++
-		if emptyMessagesCount > stream.emptyMessagesLimit {
-			err = ErrTooManyEmptyStreamMessages
-			return
-		}
-
-		goto waitForData
-	}
-
-	line = bytes.TrimPrefix(line, headerData)
-	if string(line) == "[DONE]" {
-		stream.isFinished = true
-		err = io.EOF
 		return
 	}
 
@@ -70,7 +26,7 @@ waitForData:
 }
 
 func (stream *CompletionStream) Close() {
-	stream.response.Body.Close()
+	stream.streamReader.Close()
 }
 
 // CreateCompletionStream — API call to create a completion w/ streaming
@@ -93,11 +49,12 @@ func (c *Client) CreateCompletionStream(
 	}
 
 	stream = &CompletionStream{
-		emptyMessagesLimit: c.config.EmptyMessagesLimit,
-
-		reader:         bufio.NewReader(resp.Body),
-		response:       resp,
-		errAccumulator: newErrorAccumulator(),
+		streamReader: &streamReader{
+			emptyMessagesLimit: c.config.EmptyMessagesLimit,
+			reader:             bufio.NewReader(resp.Body),
+			response:           resp,
+			errAccumulator:     newErrorAccumulator(),
+		},
 	}
 	return
 }
